move test to gpu instance
